requirements.txt
# Core backend & frontend
fastapi
uvicorn
streamlit
pandas
scikit-learn
sentence-transformers
huggingface_hub
requests
beautifulsoup4
selenium
faiss-cpu
langchain
langchain-community
google-generativeai
python-dotenv
openai